Add in custom root_patterns.
The default `root_patterns()` function that nvim-lsp provides
is great, but for Scala there will be times where the user
is in a situation where they have nested build files for a
workspace. By default this won't work since the client will
re-initialize causing the root workspace to be wrong. This
custom `root_patterns()` function adds in an extra check to
make sure the parent directory doesn't also have a build file.