Add in ability to enable metals/status
By default `statusBarProvider` is set to showMessage.
This allows a user to override this and set it to `on`,
which enables `metals/status`. Then you can use the
newly created `metals_status` function to answer the callback
for `metals/status`, and also access the `metals_status`
global variable to use in your statusline.