Add in support for settings.
This adds in the support to send settings via `workspace/didChangeConfiguration`.
At the moment this is only sent right after the initialize. This also adds in a default
empty config table for the user to use if they are editing any settings. They can
require this like so:

```lua
require'metals'.bare_config
```

The logic internally to handle merging tables is also improved which makes
the way all of the config values are created much simpler.

All docs are also updated with the new examples and all available settings.