Trim down readme and point to help docs.
One thing I did wrong with coc-metals was duplicate the stuff
I put on the website and in the readme. I don't want to make that
mistake again with having docs in various places. In this commit
I slim down the readme and point people to the help docs and also
to the example configuration in the discussions. I would like to
keep the readme as thin as possible while keeping the help docs
as rich as possible and the center of truth. This will bring the
most benifit since you can both view them on github and also
while in neovim.