Refactor log.lua.
There were large parts of log.lua that weren't being used, and I wanted to be able
to log and show something with one call, but not all the time. This introduced a new
log api that allows you to do log.info() and also provides a log.info_and_show() which
will both log and show the user. The same for warning and error.

I also replaced all usages of print with this in various places. This may be a bit excessive
in logging, but we'll see.