Refactor UI module.
Instead of the semi-transparent window that is hard to see and looks
ugly, replace the UI module with one that creates windows based on
content size, but still centered in the screen. Plus, it adds a nice
border and a title to that border. This will hopefully unlock other
better UI stuff with windows.