Add in a g:metals_use_global_executable option.
Up until this point nvim-metals fully handled the Metals install and
updates, and there was no option to _not_ have it do so. By adding in
this setting users can now just set `g:metals_use_global_executable` to
true, and then nvim-metals will default to using whichever Metals
executable is found on the $PATH.

NOTE: That this isn't really recommended to use as most of the time it's
preferable to just let nvim-metals handle this, but if you're in the
group of Nix users or someone that wants to use the same executable for
multiple different clients, this can be useful to you.