Ensure nvim-metals cache dir is created right away.
If not in the situaton where a user open a Scala project for the very
first time without the cache dir every being created, it will throw a
warning. This is because we are trying to log before we've installed
Metals which is when the cache dir was created before.