Add in ability to restart server.
The original idea behind this is for the workflow I have of:
  1. publishing Metals
  2. re-bootstapping Metals with `:MetalsInstall`
  3. Wanting to then test those changes without having to close and
     re-open nvim.

This adds in a new `:MetalsRestartServer` which will allow for me to do
just that. Also, it adds in a `:MetalsStartServer` in the case that a
user is using this for the first time and installs Metals. To get
started they used to have to close and reopen, but now after the initial
install they can just use the new `:MetalsStartServer` command.