Remove `:Format` command.
This was sort of left over from when I just started the plugin and was
also creating extra LSP helper commands. This is the last one left that
I sort of just forgot about, and it probably just shouldn't be in here.
All it's doing is just calling `vim.lsp.buf.formatting()`, so it can
very easily be mapped locally. Plus most people may already have this
mapped directly calling that anyways.

Plus, is is also conflicting with another plugin:
  https://github.com/mhartington/formatter.nvim/issues/35