Remove our `textDocument/hover` implementation.
The main reason we had our own `textDocument/hover` in here was due to the
default one not wrapping. This caused some issues with type signatures that were
quite long and not being able to see the whole thing. Now due to the improvements
in https://github.com/neovim/neovim/pull/14649 we can just remove our since the
default one now wraps.