Add in ability to toggle boolean settings.
This will allow for a user to easily use a mapping to toggle a boolean
setting. Thinking about this a bit more, we can probably in the future
make a nice way to change any setting value and send that in. Something
that prompts the user for input, and then they can input the new value
and the same logic could be used to send in the notification to Metals.
But until we need that, this should be fine.