Add in ability to specify serverProperties and also grab javaopts.
Unlike some of the other Metals clients, up until this point it's been
pretty much impossible to set server properties for Metals. This change
adds in a new `serverProperties` setting that allows you to send in
properties that Metals will be started with.

This also uses https://github.com/ckipp01/nvim-jvmopts to grab the
values from the workspace `.jvmopts` file and also values from
`JAVA_OPTS` and `JAVA_FLAGS`. However, nvim-metals strips out most of
the settings from these if they are related to memory. Mainly because
the memory requirements you put in your workspace are more than likely
for sbt, not for metals. Instead use `serverProperties` for this. I do
this to align with coc-metals and the VS Code extension.