Revert commit about removing -F recommendation.
So the main reason we need this is because nvim-metals adds in some extra
logging and messages to help users get started. So for example say you try
to get started but haven't actually installed Metals, it will log and give instructions
to you about how to do that. However, without removing -F, this won't happen.
So you can remove this if you're confident you don't need it and everything
will work fine, you just might not get messages about things going wrong