Fix handler compat for requests.
I don't fully understand why the compat function won't work
as expected for requests that expect something back. So this
fixes the issue, but I feel like it shouldn't be necessary. I'll
continue to look into this to see if there is somethign else I can do
to better align this, but for now this will fix it.

Close #217