Simplify config cache and make bare_config a function.
*NOTE* This commit contains a breaking change.

Previously when you were setting up your config table to pass into
`initialize_or_attach()` you'd start by creating a bare config like
this:

```lua
Metals_config = require("metals").bare_config
```

The problem with this was that the `bare_config` was just a table that
by you setting values to was actually mutating the `bare_config` table
internally. Since we'd like to keep that as a fresh bare config this
has been migrated to instead be a function that returns a bare config.
This means that where you are using `bare_config` you'll now want to
call it like so:

```lua
Metals_config = require("metals").bare_config()
```

Also, please subscribe to this discussion:
https://github.com/scalameta/nvim-metals/discussions/253

All breaking changes will be listed there.