fix: better check for found pattern
This fixes the case where it finds the root pattern in a directory, but
then the parent also has a different pattern that matches, which then
causes it to believe that is the root when it shouldn't be. This happens
in the cases where the root is one layer up from the .git dir.