Correct packer.nvim syntax
In lua there are no named arguments, this is simulated by the fact that when passing a table or string literal, the parenthesis can be omitted. `use` expects a table with its first value being the direction of the extension, and among other optional fields, a table in the key `requires` with the directions of the extensions required. 

In the notes, it shows as if `use` takes two arguments, a table with the direction of the first extension and then a named argument with a table with the direction of the first extension.