Add in the ability to automatically install latest snapshot.
This makes several different changes.

1. Introduces a new way to set the version to the latest snapshot. Instead of always
having to find the latest snapshot and manually set it, we instead pull it from the
maven-metadata.xml file and automaticaly use that. There has been issues in the past
with this not syncing as fast as you want, so it's not bullet-proof but should work
for most use cases. To use this just set your serverVersion as `SNAPSHOT`.

Closes #122

2. Starts work on #234. Not the preferred way to set your serverVersion or serverOrg
is to set it in the metals settings table, not as options.