feat: bump initialize_timeout_sec to 10
Thanks to https://github.com/mfussenegger/nvim-dap/pull/376 the timeout
can now be customized. The default in nvim-dap is 4 seconds, which is
just too short with Metals. This will avoid the warning being shown all
the time. This isn't configurable for the time being, but if it needs to
be in the future it can be.