refactor: ensure windows warning gets shown with others.
This just moves things around a bit. Previously we would log the windows
message, but then also again log the generic coursier message. We don't
need to do both and this makes it so we don't need to do two logs, but
rather include the windows message in the other warnings if metals or
coursier isn't found.