feat: take into account tooltip and command in meatls/status
I'm going to probably play around with this for a while to see if I like
it before merging, but the idea here is that in situations where
something goes wrong it might be hard to tell because you only get the
small status message. Now you'll actually get a select that will show
you the tooltip and the command if there is one, or it will just warn
you of what the tooltip says.

Fixes #102