feat: account for version 3 of doctor
Version 3 of the doctor format adds in some nice structured goodies in
the header instead of a big string. For example now we also get the jdk
version, the metals version, the bloop version, etc all nicely displayed
for the user. We also kept in logic to ensure that users on old metals
can still use the doctor without stuff blowing up