feat!: utilize new autocmd api
This is a breaking change. If you see error messages related to autocmds
after updating this you'll want to ensure that you're on the newly
released 0.7.0 of Neovim.
  - https://github.com/neovim/neovim/releases/tag/v0.7.0

The diff in this commit will give you a good idea of how to migrate the
old style of autocmds to the new style.

Closes #347