test: start adding in handler tests
The goal will be to test all of the custom handlers for Metals. While we
probably won't be able to test the full cycle of communication with
these between server and client we can at least test the Neovim side of
things and mock what we need to. This will at least ensure that when a
specific handler is called we aren't using an old API and we can catch
changes in the loads that Metals is sending us.

This first test covers `metals/quickpick`