refactor(decorations): migrate decorations to use extmarks
Since `nvim_buf_set_virtual_text` is deprecated I figured it was time to
do a migration to the new way of dealing with virtual text using
extmarks. In some ways it's more verbose, but there are more
possiblities. This also slightly changes the behavior now with the
virtual text shifting with the lines until they are replaced. I'm not
sure I love that, but I'm not 100% sure there is an easy way to avoid
that unless we do a clear when entering insert mode. We'll see if people
complain.

This also adds in tests to cover the decorations package.