feat: allows toggle logs to be horizontal or vertical
This makes a couple small changes that allows the toggle logs command to
be able to take a param to either open vertically or horizontally. It
now takes a direction of either `vsp` or `sp` with a default of `vsp`.
This also makes a small change that when we create the user commands we
pass in whether or not an arg is expected, and you can set the amount of
args in the commands table.

Closes #250