refactor: ignore status about external sources not compiling
This message steals your focus and forces the user to hit enter which is
pretty annoying. I hit on this quite a bit. This first came into play
due to us adding better support for status actions, but this is one case
that just doesn't work well in nvim. So for now, we ignore it.