feat: add in metalsBinaryPath
Another feature that should be useful for Nix users. In the situation where
they don't want nvim-metals to control the install, but also don't want to
use a global executable, this will allow them to instead use `metalsBinaryPath`
which allows you to pass in any path.

Know that you can't use this _with_ `useGlobalExecutable`. If you do, it will take
priority and win.

Addresses #454