fix: remove manual didChangeConfiguration call
This is now done in core via https://github.com/neovim/neovim/pull/18847
so we can safely remove this from nvim-metals. I also realized that this
was probably breaking assumptions in the past if anyone tried to
override the `on_init` handler. Now this should work as expected when a
user tries to override.

Closes #465