refactor: don't use file URI in decoded buffer
Instead of just using the fileUri instead of the name we give the buffer
while using the decoder feature we instead can use the requestedUri that
is returned with the decoded response. This is unique for each class in
a file and for each build target. This way you can now view 2 different
classes in the same file with no issue.

Fixes #459