fix: ensure close is called on BufDelete
Previously if a user would close the TVP panel they would still get
`metals/treeViewDidChange` notifications sent from the server which
would cause it to pop open again even though the user had it closed.
This ensures that a proper `metals/treeViewVisibilityDidChange` is sent
to let Metals know it was closed.

Fixes #489