refactor: change toggle_logs to open a new tab
I recently used `:LspLog` and really liked how it opened in a new tab. I
feel like this is the way to go instead of trying to do a `sp` or `vsp`.
This pr changes `toggle_logs` to now no longer take a direction, but to
instead just open a new tab.