refactor: use the actual root in config cache for toggle logs
Instead of relying on `.metals/metals.log` we now grab the actual root
uri from the config cache and use that to put together the location that
the logs file should be at. This will ensure that no matter where you
toggle the logs from or where you open the project from that we are
always looking in the right place.

Fixes: https://github.com/scalameta/nvim-metals/issues/530