refactor: make error for null in reveal in tree better
This will hopefully give the user a bit more information about what is
happening when they receive a `null` from the tree view. Since it's not
supported yet in Scala 3 this may happen. So in that situation we tell
them that and also let them know that if they are in Scala 2 they should
create an issue. We also better catch the situation where they truly get
a view id that isn't supported. In that case we show them what id it is.

Closes #546