fix: don't consider / as a valid root
There are situations where a user might open a standalone scala file
which will be a valid scala-cli project. If they do this nvim-metals
will traverse down trying to find a root pattern, and it will search all
the way to / where it _may_ find a `~/.bloop` dir and think that's the
root... We don't want that, so we don't allow `/` to be a root.