fix: make sure new lines are taken out of the short summary
In the new doctor short summary there is the potential to be new lines.
When adding things into the output of the doctor it was blowing up. This
ensures that these are stripped out and instead the multiple lines are
turned into a table and then added to the doctor output.