fix: check for the existence of hoverMessage before using
In theory the way we use decorations (not being an inline decoration
provider) we shouldn't hit on this and never have. However
https://github.com/scalameta/metals/issues/5843 caused me to realize
that `hoverMessage` is optional and we shouldn't assume it will always
be there. This change checks the existence of it before actually trying
to store the message to display later.