refactor: no longer show the statustips
This is causing all sorts of issues since it's appearing all the time
with messages from the BSP server or from Metals that really aren't that
useful and are often just telling the user to open the doctor. In VS
Code this isn't an issue since the pop up doesn't steal your focus and
just goes away by itself, but that's not possible to have an actionable
popup (as far as I know) in nvim, so instead we're going to just log
them.