refactor: remove Metals version info from Info
Instead of displaying this here, you should instead use the Metals
doctor to display this. This will remove confusion about the "supported
scala versions" that Metals dispalys which may not out of date when we
add in support for new metals version in snapshots but don't do a
release causing the help to not be updated.