fix: ignore responses from other clients
There was a change a while back that makes the buf request go to all the
attached clients. For Scala normally this isn't an issue at all, but if
you're using co-pilot which is also a client in this situation it will
fire metals requests to it which you don't want. We then throw an error
in that situation while the command still worked. This is confusing for
users, so this should fix the issue by filtering out the responses that
don't match metals.